rename modal component to be more specifig
